改单时，新生成的联系单，状态为审核完成，不再需要审核；

+++ src/main/java/com/eservice/api/web/ContactFormController.java
@@ -305,7 +305,7 @@ public class ContactFormController {
             }
             //更新 联系单的审核记录
             contactSign.setUpdateTime(new Date());
-            if(contactSign.getContactFormId() != contactForm.getId()){
+            if(contactSign.getContactFormId().intValue() != contactForm.getId()){
                 message = " contactSign里的ContactFormId 和 联系单的id 不匹配！";
                 throw new RuntimeException();
             }
